Renamed and removed some files I don't need my old ForwardRingList anymore
