+++ petrov.nikita/S1/main.cpp
@@ -1,4 +1,4 @@
-#include "forward_ring_list_2.hpp"
+#include "forward_ring_list.hpp"
 #include <string>
 #include <iostream>
 #include <limits>
